Minor refactor (#1957)
* moving codes around

* move some evm helpers

* fix

* update

* update

* fix
